Redo digit handling. -N, +N, binary, octal, hex, and exponents. (Fixes #9)
Integers can now be assigned any signed 64-bit integer value. Integers can now
be octal (0c...), hex (0x...), binary (0b...), as well as the standard
decimal.

Decimal exponents are denoted by e, and can be negative (1e-1), or positive
(1e+1 or 1e1). Numbers can also be denoted by .1 instead of 0.1 now.

All integer values can be explicitly positive and negative now (ex: -0b1
and +0c77 are allowed). This required a change in sanity, where an old test
was doing n-2 and the scanner was seeing n and -2 instead of n, -, and 2.

I also took out leftover fprintf's from the last commit.

Finally, lily_value was changed to reflect that integers are 64 bits
instead of just being 'int'.

+++ lily_syminfo.h
@@ -22,7 +22,7 @@
 /* This is where the raw data gets stored. Anything not an integer or a number
    is stored in ptr and cast appropriately. */
 typedef union {
-    int integer;
+    int64_t integer;
     double number;
     struct lily_str_val_t *str;
     struct lily_method_val_t *method;
